refactor SVGElement out of StateChart
move brace imports to ToolPanel (from StateChart)
change react-ace mode from JavaScript to TypeScript
remove unnecessary import from toMachine test
refactor SVG-related code from StateChart component into a new component
